Finalização do projeto

+++ src/infra/entities/tickers.py
@@ -1,9 +1,8 @@
 from infra.configs.base import Base
-from sqlalchemy import String, Integer, Numeric, DateTime, Numeric
+from sqlalchemy import String, Integer, Float, DateTime
 
 def get_depara_database_tickers():
     return {
-
             "Conta":Integer,
             "DataCompra":DateTime,
             "Lancamento":String,
@@ -13,16 +12,16 @@ def get_depara_database_tickers():
             "QuantidadeMovimentacao":Integer,
             "SubMercado":String,
             "Ticker":String,
-            "ValorMovimentacao":Numeric,
+            "ValorMovimentacao":Float,
             "TickerItem":Integer,
             "QuantidadeLa":Integer,
             "QuantidadeFinal":Integer,
-            "EstoqueLa":Numeric,
-            "EstoqueMais":Numeric,
-            "EstoqueMenos":Numeric,
-            "DeltaEstoque":Numeric,
-            "EstoqueFinal":Numeric,
-            "PrecoMedio":Numeric
+            "EstoqueLa":Float,
+            "EstoqueMais":Float,
+            "EstoqueMenos":Float,
+            "DeltaEstoque":Float,
+            "EstoqueFinal":Float,
+            "PrecoMedio":Float
 
         }
